Split SDEServices in to SDEService and SDEServiceManager and simplify the Import function.
