+++ src/WHMapper/Services/SDE/SDEConstants.cs
@@ -1,0 +1,9 @@
+﻿namespace WHMapper.Services.SDE
+{
+    public static class SDEConstants
+	{
+        public const string REDIS_SOLAR_SYSTEM_JUMPS_KEY = "solarysystemjumps:list";
+        public const string REDIS_SDE_SOLAR_SYSTEMS_KEY = "solarsystems:list";
+    }
+}
+
