+++ src/WHMapper/WHMapper.csproj
@@ -130,7 +130,6 @@
     <Folder Include="Models\DTO\Anoik\" />
     <Folder Include="Services\EveAPI\User Interface\" />
     <Folder Include="Resources\SDE\" />
-    <Folder Include="Services\SDE\" />
     <Folder Include="Models\DTO\SDE\" />
     <Folder Include="Services\EveMapper\" />
     <Folder Include="Services\EveMapper\AuthorizationPolicies\" />
